+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Angular UI Changes
 ===============
 
-### Release 0.1, In Progress
+## Release 0.1.0, Done
 
-- Created Transition.
+### 2014/01/14
+- Created Buttons module. Module contains two directives that can make a group of buttons to behave like a set of checkboxes or radio buttons.
+
+### 2014/01/11
+- Created Position class.